refactoring AppDataBase.kt
