+++ app/src/main/java/com/w4ereT1ckRtB1tch/moviefan/di/DataBaseModule.kt
@@ -3,7 +3,6 @@ package com.w4ereT1ckRtB1tch.moviefan.di
 import android.content.Context
 import androidx.room.Room
 import com.w4ereT1ckRtB1tch.moviefan.data.db.AppDataBase
-import com.w4ereT1ckRtB1tch.moviefan.data.db.FilmDataBaseScheme
 import com.w4ereT1ckRtB1tch.moviefan.data.db.RepositoryMock
 import dagger.Module
 import dagger.Provides
@@ -19,7 +18,7 @@ object DataBaseModule {
         Room.databaseBuilder(
             context,
             AppDataBase::class.java,
-            FilmDataBaseScheme.DataBase.NAME
+            AppDataBase.NAME
         ).build()
 
     @Provides
