+++ app/src/main/java/com/w4ereT1ckRtB1tch/moviefan/presentation/adapters/FooterStateAdapter.kt
@@ -1,4 +1,4 @@
-package com.w4ereT1ckRtB1tch.moviefan.presentation.recycler_adapters
+package com.w4ereT1ckRtB1tch.moviefan.presentation.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
